requirements.txt
annotated-types==0.7.0
anthropic==0.100.0
anyio==4.13.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==48.0.0
distro==1.9.0
docstring_parser==0.18.0
feedparser==6.0.12
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.196.0
google-auth==2.52.0
google-auth-httplib2==0.4.0
google-generativeai==0.8.6
googleapis-common-protos==1.75.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.13
iniconfig==2.3.0
jiter==0.14.0
MarkupSafe==3.0.3
packaging==26.2
pluggy==1.6.0
praw==7.8.1
prawcore==2.4.0
proto-plus==1.28.0
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.4
pydantic_core==2.46.4
Pygments==2.20.0
pyparsing==3.3.2
pytest==9.0.3
python-dotenv==1.2.2
python-http-client==3.3.7
requests==2.33.1
sendgrid==6.12.5
sgmllib3k==1.0.0
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
update-checker==0.18.0
uritemplate==4.2.0
urllib3==2.7.0
websocket-client==1.9.0
Werkzeug==3.1.8
